Safari-611.1.10/LayoutTests/loader/navigation-policy/should-open-external-urls/main-frame-without-flag-programatic-expected.txt

 - decidePolicyForNavigationAction
<NSURLRequest URL resources/programatically-navigate-to-notify-done.html, main document URL resources/programatically-navigate-to-notify-done.html, http method GET> is main frame - yes should open URLs externally - no
 - decidePolicyForNavigationAction
<NSURLRequest URL notify-done.html, main document URL notify-done.html, http method GET> is main frame - yes should open URLs externally - no
